figure: Add Figure 8
Add all code required to reproduce Figure 8 of this study.

Note: WILL NOT WORK. ADDING TEMPLATES AND GMT METAFILES LATER.

Additionally, I have not yet finished writing a guide detailing how the
inputs for the stress model are made and how Coulomb is used to perform
the stress modeling. These will be added at a later date, but for now I
have just included the final results that are plotted.